# Removed allocations for timestamp