@nblumhardt 2024-08-02
Reading the code again, I see that we should fall through and still create correct output in the "too long" case - think it just needs a test :-)